0113 ps

+++ src/Implementation/Main_1614_영식이의손가락.java
@@ -5,15 +5,15 @@ import java.io.IOException;
 import java.io.InputStreamReader;
 
 public class Main_1614_영식이의손가락 {
-	static final int CYCLE = 8;
+	static final long CYCLE = 8;
 
 	public static void main(String[] args) throws IOException {
         BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
         
         int hurt = Integer.parseInt(br.readLine());
-        int count = Integer.parseInt(br.readLine());
+        long count = Integer.parseInt(br.readLine());
         
-        int result = 0;
+        long result = 0;
         switch(hurt) {
 	        case 1: {
         		result = CYCLE * count;
